# api_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "The Third Voice AI",
        "description": "AI-powered communication assistant for difficult conversations - Healing families through better communication",
        "version": "1.0.0"
    },
    "paths": {
        "/api/contacts/": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get Contacts",
                "description": "Get all contacts for the current user\n\nReturns list of contacts ordered by most recently updated",
                "operationId": "get_contacts_api_contacts__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ContactResponse"
                                    },
                                    "type": "array",
                                    "title": "Response Get Contacts Api Contacts  Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create Contact",
                "description": "Create a new contact\n\nDemo users are limited to a maximum number of contacts",
                "operationId": "create_contact_api_contacts__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
        "/api/feedback/": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Submit Feedback",
                "description": "Submit user feedback\n\nRatings are from 1-5 stars with optional text feedback",
                "operationId": "submit_feedback_api_feedback__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Name",
                        "description": "Contact name"
                    },
                    "context": {
                        "$ref": "#/components/schemas/ContextType",
                        "description": "Relationship context"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "context"
                ],
                "title": "ContactCreate"
            },
            "ContactResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "context": {
                        "$ref": "#/components/schemas/ContextType"
                    },
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "message_count": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message Count",
                        "description": "Total messages for this contact"
                    },
                    "last_message_date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Message Date",
                        "description": "Date of last message"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "created_at",
                    "name",
                    "context",
                    "user_id"
                ],
                "title": "ContactResponse"
            },
...
            "ContextType": {
                "type": "string",
                "enum": [
                    "romantic",
                    "coparenting",
                    "workplace",
                    "family",
                    "friend"
                ],
                "title": "ContextType"
            },
            "FeedbackCreate": {
                "properties": {
                    "rating": {
                        "type": "integer",
                        "maximum": 5.0,
                        "minimum": 1.0,
                        "title": "Rating",
                        "description": "Rating from 1-5 stars"
                    },
                    "feedback_text": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1000
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Feedback Text",
                        "description": "Optional feedback text"
                    },
                    "feature_context": {
                        "type": "string",
                        "title": "Feature Context",
                        "description": "Which feature this feedback is for"
                    }
                },
                "type": "object",
                "required": [
                    "rating",
                    "feature_context"
                ],
                "title": "FeedbackCreate"
            },
            "FeedbackResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "rating": {
                        "type": "integer",
                        "title": "Rating"
                    },
                    "feedback_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Feedback Text"
                    },
                    "feature_context": {
                        "type": "string",
                        "title": "Feature Context"
                    },
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "created_at",
                    "rating",
                    "feedback_text",
                    "feature_context",
                    "user_id"
                ],
                "title": "FeedbackResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}